moved some stuff around and created components for new pages
