+++ frontend/crd/src/app/milestones-page/milestones/milestones.component.ts
@@ -22,5 +22,4 @@ export class MilestonesComponent implements OnInit {
 
   milestones: Array<Milestone> = []
 
-  protected readonly Milestone = Milestone;
 }
