+++ frontend/crd/src/app/portfolio/portfolio.component.spec.ts
@@ -2,7 +2,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { PortfolioComponent } from './portfolio.component';
 import {MockComponent} from "ng-mocks";
-import {MilestonesComponent} from "./milestones/milestones.component";
+import {MilestonesComponent} from "../milestones-page/milestones/milestones.component";
 
 describe('PortfolioComponent', () => {
   let component: PortfolioComponent;
